refactor of accountingImplmentation.py and Utils complete. Will start writing tests now
