+++ code_src/staking/dot/fxn_decorator_implementations/accountImplementation.py
@@ -10,8 +10,6 @@ from code_src.staking.dot.fxn_decorator_implementations.accountImplementationUti
 
 class AccountImplementation:
     """
-    * TODO: rename to AccountImplementation (big refactor)
-
     AccountImplementation is a class for containing functions related to accounts.
     An "account" is defined as keypair with associated data.
 
@@ -23,7 +21,7 @@ class AccountImplementation:
     Therefore, some of the functions in AccountImplementation are "redundant", e.g. 
     createMnemonic just calls MnemonicImplementation in accountImplementationUtils.py
     """
-    def __init__(self, logger, mnemonic="", ss58_address=""):
+    def __init__(self, logger, mnemonic=None, ss58_address=None):
         self.mnemonic = mnemonic
         self.ss58_address = ss58_address
         self.logger = logger
@@ -67,20 +65,31 @@ class AccountImplementation:
 
             self.logger.info(f"""account {ss58_address} infos
 
-    nonce : {value['nonce']}
-    consumers : {value['consumers']}
-    providers : {value['providers']}
-    sufficients : {value['sufficients']}
-    free : {free} {activeConfig.coinName}
-    reserved : {reserved} {activeConfig.coinName}
-    misc_frozen : {misc_frozen} {activeConfig.coinName}
-    fee_frozen : {fee_frozen} {activeConfig.coinName}
-    """)
+            nonce : {value['nonce']}
+            consumers : {value['consumers']}
+            providers : {value['providers']}
+            sufficients : {value['sufficients']}
+            free : {free} {activeConfig.coinName}
+            reserved : {reserved} {activeConfig.coinName}
+            misc_frozen : {misc_frozen} {activeConfig.coinName}
+            fee_frozen : {fee_frozen} {activeConfig.coinName}
+            """)
         except Exception as e:
             self.logger.error(f"{e}")
 
-#    def getAccountBalanceForBonding(self):
-#    def getAccountBalance():
+    def getAccountBalance(self, purpose=None):
+        """
+        TODO: Document
+        """
+        if(purpose == None):
+            # TODO: improve this to return dictionary with account values
+            self.getAllAccountInfo()
+        elif (purpose == "bonding"):
+            return AccountBalanceForBonding(self).getAccountBalanceForBonding()
+        else: # "undefined" scneario - error
+            self.logger.warning(f"Unknown object passed into getAccountBalance. Failing.")
+            sys.exit(0)
+
 
 class DotAccountCall:
     """
@@ -115,19 +124,25 @@ class AccountBalanceForBonding:
     Class is to return balance of bonding - verifications
     TODO: Better explanation
     # TODO:
-    # * rename AccountBalance class; better inheretence/abstraction
-    # * use accountImplementation
-    # * called in substrateCallImplementationUtils.py
-    # Why doens't this go in DotAccountCall?
+    # * Ideally keep in Utils file instead of here, but need to refactor 
+    # to prevent importing accountImplementation.py in Utils
     """
-    def __init__(self, logger, ss58_address, account: AccountImplementation):
-        self.logger = logger
-        self.ss58_address = ss58_address
+    def __init__(self, account: AccountImplementation):
+        self.logger = account.logger
+        self.ss58_address = account.ss58_address
 
         # takes AccountImplementation as initialization arguement
-        if(isinstance(account, AccountImplementation()) == False):
+        if(isinstance(account, AccountImplementation) == False):
+            print("""AccountImplementation type *not* passed to initialize AccountBalanceForBonding.\n
+                Therefore, the logger is not set and this erorr is printed instead of logger.
+                Program Exiting...""")
+            # TODO: If wrong object passed, logger will not be set and not print the error. Fix.
             logger.warning("AccountImplementation type *not* passed to initialize AccountBalanceForBonding. Failing.")
+            sys.exit(0)
             return False
+
+    def __call__(self):
+        return self.getAccountBalanceForBonding()
 
     def getAccountBalanceForBonding(self):
         # query balance info for an account
@@ -143,107 +158,6 @@ class AccountBalanceForBonding:
         totalAccountBalance = free / activeConfig.coinDecimalPlaces + reserved / activeConfig.coinDecimalPlaces
 
         return totalAccountBalance
-
-class MnemonicImplementation:
-    """
-    Class creates a mnemonic and prints in the log, currently has no other purpose
-    * For security reasons, do not store the mnemonics
-    * This class is intentionally separate from AccountImplementation as there may be times
-    when features of mnemonics should be added/changed without concerning AccountImplementation
-    """
-    def __init__(self, logger):
-        self.logger = logger
-
-    def createMnemonic(self):
-        mnemonic = Keypair.generate_mnemonic()
-        createMnemonicLogMessage = f"""create mnemonic phrase
-
-    mnemonic : {mnemonic}
-
-    note : please write down this mnemonic in paper and stored in a save place.
-    learn more about mnemonic : https://coinmarketcap.com/alexandria/glossary/mnemonic-phrase
-            """
-
-        try:
-            self.logger.info(createMnemonicLogMessage)
-            return mnemonic
-        except Exception as e:
-            self.logger.critical(f"error : {e}")
-            return False
-
-class KeyPairImplementation:
-    """
-    Class creates a keypair
-    """
-    def __init__(self, logger, mnemonic):
-        self.logger = logger
-        self.mnemonic = mnemonic
-        # what to do if no mnemonic is passed? Adapt fxn signature.
-
-    # previously dotCreateKeyPair
-    # def createKeyPair(self):
-    def getAddressFromMnemonic(self):
-        """
-        Calculates the dot address given a mnemonic and prints and returns it (or exits the system if it fails).
-        It's currently kept outside the DotAccountCall as an auxilary function in order to keep the pre-defined
-        function set in DotAccountCall (i.e. createMnemonic, getAccountInfos, etc.)
-        Function is called from DotAccountCall and DotSubstrateCall
-        """
-        invalidCharacters = "[@_!#$%^&*()<>?/|}{~:]0123456789"
-
-        # If a mnemonic is not passed in, the default in the above library will be used
-        # however, we will enforce that "something" is passed in to avoid the default (len 10 is arbitrary)
-        if (len(self.mnemonic) < 10):
-                logger.critical("A bad mnemonic as been passed to create the keypair")
-                return False
-
-        try:
-            # Keypair ~ https://github.com/polkascan/py-substrate-interface#keypair-creation-and-signing
-            key = Keypair.create_from_mnemonic(mnemonic=self.mnemonic, ss58_format=activeConfig.ss58_format)
-            self.logger.info(f"""Here is the address associated with the above mnemonic:\n
-        {key}
-         \n\n""")
-
-            # do a quick verification that the key signs normally
-            if key.verify("This is a test message", key.sign("This is a test message")):
-                return key
-            else:
-                # if the key verification fails, exit immediatly
-                self.logger.critical("\nDO NOT USE KEY. KEY INCORRECTLY GENERATED.\n")
-                return False
-
-        except ValueError:
-            # more thorough check for the mnemonic below
-
-            # split mnemonic by space into words
-            splitMnemonic = self.mnemonic.split(" ")
-
-            lengthMnemonic = len(splitMnemonic)
-            # check word length and special character
-            lengthWordInMnemonic = any(word for word in splitMnemonic if len(word) < 3 or len(word) > 8)
-            lengthOfDigitInMnemonicIfAny = any(s for s in self.mnemonic if s in invalidCharacters)
-
-            # Checking mnemonic length
-            # length doesn't meet the standard
-            if lengthMnemonic not in [12, 15, 18, 21, 24]:
-                self.logger.critical(
-                    "Number of words must be one of the following: [12, 15, 18, 21, 24], but it is not (%d)."
-                    % lengthMnemonic)
-                return False
-
-            # length meet the standard
-            else:
-                # Checking mnemonic for invalid characters (non alphabet)
-                if lengthOfDigitInMnemonicIfAny:
-                    self.logger.critical("Mnemonic words must be alphabet words.")
-                    return False
-
-                # check word len in mnemonic min = 2 max = 8 or the mnemonic doesn't have valid word
-                elif lengthWordInMnemonic or not bip39_validate(self.mnemonic):
-                    self.logger.critical("Please check for messing strings.")
-                    return False
-
-
 
             ### TEST THIS CLASS BEFORE DELETING BELOW ###
