+++ code_src/staking/dot/fxn_decorator_implementations/substrateCallImplementation.py
@@ -7,10 +7,7 @@ from substrateinterface.exceptions import SubstrateRequestException
 from config import activeConfig
 from Logger import myLogger
 from code_src.staking.dot.fxn_decorator_implementations.substrateCallImplementationUtils import *
-# TODO: 
-# consider adding dotCreateKeyPair() to DotAccountCall so that this file only has to import
-# accountImplementation and not a utils file for accountImplementation... importing another utils file isn't very clean
-from code_src.staking.dot.fxn_decorator_implementations.accountImplementationUtils import *
+from code_src.staking.dot.fxn_decorator_implementations.accountImplementation import *
 
 
 class DotSubstrateCall:
@@ -57,9 +54,9 @@ class DotSubstrateCall:
         #this_keypair = dotCreateKeyPair(logger=self.logger, mnemonic=self.seed)
         # TODO: should call AccountImplementation().createAccount() instead to keep everything needed in accountImplementation.py
         # this_keypair = KeyPairImplementation().getAddressFromMnemonic()
-        this_keypair = accountImplementation(logger=self.logger, mnemonic=self.seed)
+        this_address = AccountImplementation(logger=self.logger, mnemonic=self.seed).getAddressFromMnemonic()
 
-        extrinsic = activeConfig.activeSubstrate.create_signed_extrinsic(call=call, keypair=this_keypair)
+        extrinsic = activeConfig.activeSubstrate.create_signed_extrinsic(call=call, keypair=this_address)
         """None
         {'substrate': < substrateinterface.base.SubstrateInterfaceobject at 0x000001F7C2F73F70 >, 
         'extrinsic_hash': '0xa9719ca1430e4a9f0305b03e4b6bdd582458525bcc44b01db3caa7fa7d933867',
